提升为u64，可兼容2T以上大硬盘，并修复遗留bug

+++ modules/iopcore/cdvdman/device-bdm.c
@@ -131,7 +131,7 @@ void DeviceUnmount(void)
     DPRINTF("%s\n", __func__);
 }
 
-int DeviceReadSectors(u32 lsn, void *buffer, unsigned int sectors)
+int DeviceReadSectors(u64 lsn, void *buffer, unsigned int sectors)
 {
     int rv = SCECdErNO;
 
@@ -141,7 +141,7 @@ int DeviceReadSectors(u32 lsn, void *buffer, unsigned int sectors)
         return SCECdErTRMOPN;
 
     WaitSema(bdm_io_sema);
-    if (bd_defrag(g_bd, cdvdman_settings.fragfile[0].frag_count, &cdvdman_settings.frags[cdvdman_settings.fragfile[0].frag_start], ((u64)lsn) * 4, buffer, sectors * 4) != (sectors * 4))
+    if (bd_defrag(g_bd, cdvdman_settings.fragfile[0].frag_count, &cdvdman_settings.frags[cdvdman_settings.fragfile[0].frag_start], lsn * 4, buffer, sectors * 4) != (sectors * 4))
         rv = SCECdErREAD;
     SignalSema(bdm_io_sema);
 
@@ -152,20 +152,20 @@ int DeviceReadSectors(u32 lsn, void *buffer, unsigned int sectors)
 // oplutils exported function, used by MCEMU
 //
 
-void bdm_readSector(u32 lba, unsigned short int nsectors, unsigned char *buffer)
+void bdm_readSector(u64 lba, unsigned short int nsectors, unsigned char *buffer)
 {
     DPRINTF("%s\n", __func__);
 
     WaitSema(bdm_io_sema);
-    g_bd->read(g_bd, (u64)lba, buffer, nsectors);
+    g_bd->read(g_bd, lba, buffer, nsectors);
     SignalSema(bdm_io_sema);
 }
 
-void bdm_writeSector(u32 lba, unsigned short int nsectors, const unsigned char *buffer)
+void bdm_writeSector(u64 lba, unsigned short int nsectors, const unsigned char *buffer)
 {
     DPRINTF("%s\n", __func__);
 
     WaitSema(bdm_io_sema);
-    g_bd->write(g_bd, (u64)lba, buffer, nsectors);
+    g_bd->write(g_bd, lba, buffer, nsectors);
     SignalSema(bdm_io_sema);
 }
